Add meta.dandiset.identifier index
Add an index for the get_dandiset endpoint. A compound index on
`meta.dandiset.identifier` and `parentId` is used to ensure that the
get_dandiset endpoint uses the correct index. `meta.dandiset.identifier`
is the first key for the index so that an index on that field is
available in the future.