Parameterize test authentication
Currently we only test server requests with a user, which corresponds to
a USER_AUTH level token. The Dandi CLI uses a token with only DATA_READ
and DATA_WRITE permissions, which is a lower permission level.
Parameterize our API tests so they test both users and read/write
tokens.