RF: nwb.json => dandiset_metanwb.json
Initially I had missed the fact that it is not the schema we impose on top
of nwb, but rather a schema for dandiset.yaml fields which are derived from
metadata we find in nwb files.  As it is not directly the information from
nwb, I decided to call it metanwb.

This is a follow up to previous renamings in bdf09cfed3d76a7dc73de7bcf82bdd28f984c9f2