Make owner permission operations recursive
This has a notable side effect, which is if a user has existing
permissions on a subfolder (but not on the top-level dandiset folder),
the permissions of that user will be overwritten based on any owner
operation done to the root dandiset folder. If they are included in an
owner addition, they will be upgraded to admin on all subfolders. If
they are not included, all of their permissions will be removed from all
subfolders.