Fetch version info from publish API
Define a new axios client for using the publish API.

Define a new `publish` store for storing stuff from the publish API,
much as we have a `girder` store to store stuff from girder.

Display the published versions on the DandisetLandingView. This display
code is temporary to prove that the underlying code works as intended.