Fix e2e bug
Fix a bug where the Dandiset list failed to render during testing.

If the publish service is unreachable, an error was thrown which
interrupted rendering of the list view. During e2e testing the publish
service is not run, so this fails all e2e tests.

The solution is to check for `Network Error`s and handle them gracefully
by assuming there is no published version.