Add sync to immediately before publish
Despite best intentions, syncing dandisets during dandiset creation
makes things much more complicated. If the draft dandiset is created,
but the sync fails, then there is either an orphaned draft dandiset that
needs to be deleted, or the two systems are desynchronized and need to
be synced up again at a later date.

Test environments would need to mock the dandi-publish service to use
the create dandiset button. This is fine in pytest, but is much harder
in the end to end testing environment. There, an instance of
dandi-publish needs to be stood up from scratch in CI, or the entire
service needs to be mocked somehow.

The easier solution is to simply sync dandisets immediately before
publish. Syncs are idempotent, so there's no problem with syncing an
existing dandiset. If the sync fails, girder can simply fail the whole
publish operation.