Add feature toggles to web client
 * Add a new feature toggle, `UNIFIED_API`, which defaults to `false`.
 * A constant is now exposed on every component for every defined
feature toggle. Components can access the new toggle with
`if (this.UNIFIED_API) {...}` or in templates with `v-if="UNIFIED_API"`.
 * Add two functions to the global scope, `enable` and `disable`.
Typing `enable('UNIFIED_API')` into the web console will enable the
feature toggle for the rest of the session, while
`disable('UNIFIED_API')` will disable it.
 * Toggles are listed in `web/src/featureToggle.js`. For development,
they can be temporarily enabled/disabled there.

Also, a minor linting change was applied to `DownloadDialog`.