Fix bug in _get_dandiset_for_item
The bug in this function didn't check that an uploaded item was located
under a collection (not a user folder, for a counter example), or that
the collection that is does reside under is in fact the dandiset drafts
collection that we care about. This resulting in errors being thrown
during item upload, when trying to reference the "dandiset" metadata
field, since it did not exist on these sub-folders.