Tweak dandiset serializer
Previously, a `MostRecentVersionSerializer` was defined to avoid
including extra information when serializing dandisets. This was more
complicated on the backend and required more special handling on the
frontend, so instead of that, Dandiset views now use a serializer that
looks like this:

```
{
  "identifier": ...
  "most_recent_version": {
    "version": ...
    ...
    "dandiset": {
      "identifier": ...
    {
  {
}
```

The dandiset identifier is now included twice, but now
`most_recent_version` uses the normal `VersionSerializer` so it doesn't
require any special handling in the client.