Switch from social_account_added to user_signed_up
It turns out that django-allauth only sends the `social_account_added`
when a new social account is connected to an existing user, not when a
new SocialAccount + User are registered. Instead we need to use the
`user_signed_up` signal and look up all of the social accounts for the
new user.