Rearrange manifest tests to reduce flake
The manifest writing tests were executing the task, opening the
generated file, then generating the expected value. The file operations
took long enough that sometimes the task ran in a different second than
the expected value generation, generating a different timestamp and
different manifest results.

The expected value was moved to immediately after running the task,
which should reduce the frequency of the problem.

This is completely unrelated to validation.