ENH: disallow non-super users creating dandisets with a specified identifier
We have got dandiset 000105 leaping from 000070.  It seems that operation was initiated
from web-ui so in principle the issue is probably not related to this change.  But especially
now that migration is over, I think it would be wise to allow for creation of dandisets
with a given identifier to be allowed only to the superusers, so someone does not
create 999999 dandiset thus making us run of the "dandiset id namespace"