Show correct last-modified time in dandiset detail panel
The non-existent `updated` property was previously being passed to
momentjs as the last modified date. Since that value was always
`undefined`, and momentjs returns the current date and time when called
without arguments (or with `undefined`, apparently), this explains why
the frontend was always showing the current time for the last modified
value.