Conver Asset.published_asset() into published_metadata() and publish()
`Asset.published_asset()` created a new `Asset`, which is not how
publishing `Asset`s should work. Instead, the `Asset` should be modified
in place, which is what `publish()` does.
`published_metadata()` is also added so that the validation task can get
the metadata of the published `Asset` without actually modifying the
`Asset`.