Remove section on pull requests and CI
It turns out that CI will run just fine from forks, so I've removed this
paragraph. The text about Netlify is informative, but not enough if you
don't know what Netlify is, and too much so if you do. There will be
a better path to documenting dev basics in the handbook, rather than in
this README.