Fix race condition when overwriting manifest file
There were some instances of Django storage's `save` method appending a
randomized suffix to the name of the checksum file before saving to
preserve the existing file contents.
I believe this was a race condition, since we were deleting the file,
then saving it. If S3 did not propogate the delete before the save,
Django would see the existing file and erroneously try to preserve it.

Use `_save`, which simply writes the file directly, instead of `save`,
which handles altering the file name.