Rename the priority-worker dyno to worker
The Girder 4 Terraform module defines two dyno types, `web` and
`worker`. You can add more types on top of that, but you cannot delete
those two.
As `priority-worker` was the primary worker, I am renaming that to
`worker` to keep Terraform happy. The alternative is ejecting completely
from that Terraform module and managing all the state ourselves, which
would be very obnoxious.