BF DOC(incomplete): /zarr/.../ingest has 204 not 200 response
Unfortunately it is no a complete fix, since for 400 there could be 2
possible responses, but only one is documented.  I did not find a way to
document both as possible valid choices, but may be it should be a different
from 400 for upload_in_progress anyways?

Also for 204 -- not sure what was motivation for using that reponse
instead of 200 with some message such as "Dispatched ingestion" to inform user
that all is good etc.