Clean up handling of missing `DOI_API_PREFIX`
This setting should never be unset in a real deployment of DANDI, so
silently replacing it with a default if it's missing is not ideal.
Instead, this changes it so that things fail loudly and an error is
raised if this setting is not supplied in an env var. In dev/testing
configurations, it is hardcoded to the placeholder value.