Fix invalid contributors causing crash
Currently, we assume that `Version.metadata['contributors']` is a list
of JSON objects. This holds true for any metadata uploaded by the DANDI
CLI, but there's nothing preventing users from uploading arbitrary JSON.
Invalid JSON will be marked as invalid, but there's some places in the
code where we assume it's valid anyway which results in a crash.