Run all linting commands
Without the `ignore_errors` configuration option set to `True`, Tox
exits as soon as any command in the list fails. In the case of the lint
tests, we would like to continue running the commands, since they are
essentially independent tests.

This was masking formatting errors whenever the `check` command was
failing, preventing us from knowing that we should fix those.