Sort by created time, descending
This fixes a bug in the emptiness detection logic. It is necessary to
look at the *newest* version in the Dandiset and count the assets there
in order to determine if a Dandiset is empty.

The bug manifested for a Dandiset that had published versions (and thus
was *not* empty) but whose draft version had been emptied of all assets.
The logic found the oldest version (which was the draft version), saw
that it was empty, and concluded that the entire Dandiset was empty.

The fix selects the newest version, which is either the draft version
(for unpublished Dandisets), or a published version. In the former case,
the emptiness of the draft version determines the emptiness of the
Dandiset; in the latter case, the Dandiset will be found to be
non-empty.