Pin the version of ruff to the 0.6.x minor version series
This avoids breakeages that can slow things down when the version of
ruff advances and gets picked up by CI, but remains at a "stable"
version in local environments. In this case, local linting/formatting
results can diverge from what happens in CI, which then causes a
scramble to make the versions equal "between dev and prod" (so to
speak).

The version specifier is `~=0.6.2` to keep the version that gets
installed to any from the 0.6.x series that is at least 0.6.2. According
to the ruff docs, patch-level bumps represent bugfixes, when minor-level
bumps are breaking changes (hence the divergence we have observed).

We avoid setting the version in pyproject.toml (using [this
setting](https://docs.astral.sh/ruff/settings/#required-version)) so
that we can put in place a CI step that *does* install the latest
version and shows us any divergences.