Add a test against the latest version of ruff available
This is a test added to the matrix of tests run by CI, but marked as
`continue-on-error` so that the full suite does not fail if this test
fails. The purpose is to signal that a later version of ruff may exist
that causes new linting/formatting failures, but allowing the dev team
to manually upgrade to the new version.