Wrap Contributors box with v-row/v-col
The `MetadataCard` wraps its contents in a `v-row` and `v-col`, which
adds some additional padding around the card. The "Contributors" card,
however, is the only DLP card that doesn't use the `MetadataCard`
component. Prior to this change, the "Contributors" card didn't use
`v-row` and `v-col`, so it didn't have the same padding as the other
cards.