Avoid v-bind for href value
This was causing a parse error on `mailto:help@dandiarchive.org` since
the v-bind directive causes the value to be interpreted as JavaScript.
In this case, the `mailto:` part looks like a label, following by
something that is not valid JS.

Since the value we want here is static, we don't need to use v-bind in
any case.