Allow null `user` when `admin` is True
This enables us to omit the user parameter for audit actions when they
occur as part of some administrative action (e.g., a bulk updating of
incorrect metadata). A null `user` is only allowed with `admin` set to
True, in which case blank values are used for the `username`, `email`,
and `fullname` parameters of the `AuditRecord` that is created.