ENH: Unify and make subject more informative on new user notification emails
I kept running into difficulty of searching up emails for a new user by github
id since it is not mentioned anywhere in the email notification to review. It
is because in one we listed  email and in another ID in subject line.

I think it would be useful to unify and make subject more informative right away but using

- the same information in the subject line
- listing important information in the subject line