Perform manual validation on full metadata
This is needed because we do not pass the full metadata to the
VJSF component, but instead break it up into smaller pieces for UI/UX
rendering purposes. Certain validation errors, such as not having at
least one `Contributor`, can't be detected unless we validate the full
metadata against the schema.