Update swagger/redocs urls to align with Resonant
Typically a Resonant app will serve the redoc and swagger UI at
/api/docs/{swagger|redoc}, but for historical reasons dandi serves them
at the root of the API. This updates the routes to match the Resonant
convention. We may want to come back to this later and deprecate/remove
the old routes after some time, but for now the old routes are still
available and simpy redirect to the new ones.