ENH: allow googlebot also /search endpoint and for filtered listing of assets
filtered listing of assets is ATM to sense if dandiset has a zarr archive in it,
so we could potentially disable publishing. There is a generic https://github.com/dandi/dandi-archive/issues/2404
asking to avoid that. But for now for DLP to load fully we need to allow for that end point.

The URL looks like
https://api.dandiarchive.org/api/dandisets/000026/versions/draft/assets/?zarr=true&page_size=1
and while reading
https://developers.google.com/search/docs/crawling-indexing/robots/robots_txt
I got impression that query options (after ?) are also considered, so we can
just overall just allow queries which are of page_size=1