Extend allowed time to fulfill PUT upload to half-hour from 10 minutes
Both numbers are arbitrary. It could be of security hazard to make them
arbitrarily long happen such URLs leak/get shared publicly.  But in both cases
durations are still reasonably short.

On some network(s) though we can run into situation that we do not upload "fast
enough" and thus running into 403s which could be (I think) avoided by allowing
for longer period of time to initiate uploads.  See

   https://github.com/dandi/dandi-cli/issues/1662

Sure think a proper solution is TODO (as good old comment says) - for client to
react and to re-request a new pull of URLs etc.  That is to be done, but I
think this PR also would make a world a better place meanwhile ;)