Switch minio docker image used by CI
Bitnami has deprecated the MinIO image we currently use in CI, and has
instituted a "brownout" today. This is the reason CI is currently
failing on `master` and new PRs. They've made the image available under
`bitnamilegacy/` for now, so this just changes it to use that until we
get switched over to whatever their new solution is.