Truncate asset validation errors/fix memory use
This `asset_validation_errors` property was evaluating the entire
QuerySet of Asset objects and building up a list of every single asset
validation error, which was causing memory issues for larger dandisets.
For a temporary fix, we truncate the list of INVALID assets such that
we only display errors for, at most, the 10 assets with the most errors
and up to 10 additional "Pending" assets. Future work will be to
reconsider the design of how we store and expose asset validation errors
in the API.