Directly add metadata instead of using `.model_construct`
The use of `Pydantic.model_construct` is causing warnings, as
the data being passed has not been validated, which violates
how the API should be used. Rather than attempt to coerce
everything to its proper type for `.model_construct` to work
(e.g. `"contributor" needs to contain a `Person`, not an
instance of `dict`), just set metadata normally.