Fix bug in publishing process
There is a bug during the publishing process where a shallow reference
to the draft `Version.metadata` dict is erroneously used for the newly
created published `Version`. This means that when we insert new fields
into the published metadata during the publishing process, we are also
unknowingly adding them into the draft `Version` object that lives in
memory. Then, during the `_publish_dandiset` service function, we call
`.save()` on the draft version, which writes those new fields to the DB.