Present the allauth logout view on clientside logout
This is the correct way to "log the user out of the API", even if it
involves users having to confirm their logout action. The reason for
that, in turn, has to do with preventing CSRF attacks against a user's
login state.

The `ACCOUNT_LOGOUT_REDIRECT_URL` governs where the user is taken after
logout completes. We can't use the logout view URL's `next` parameter
because the redirect target is on a different domain from the backend.