Move the sitemaps view to a non-canonical URL
This is to prevent crawlers from reading the sitemap directly from the
backend, since it references URLs not on the same host, which is not
allowed (see https://www.sitemaps.org/protocol.html#location).

This is in preparation for generating the sitemap server-side, but
proxying it to the frontend so that the hosts align properly.