Change how DANDI_ENABLE_SITEMAP_XML works
With this change, the backend always generates a sitemap file, but the
ENABLE setting governs whether any URL entries are included in it. This
simplifies deployment concerns: the backend always generates a sitemap,
and the frontend always proxies it to `/sitemap.xml`, but now a
deployment not wishing to advertise DLPs will simply not show any of
those URLs.