Generate robots.txt serverside (and proxy it to the frontend)
This commit restores the original robots.txt view for the backend, and
creates a new one specifically for the frontend. It's done this way to
make it easier to generate the frontend's sitemap.xml URL within
robots.txt (and it piggybacks off the similar proxying trick for
sitemap.xml itself).